# specs/001-little-teacher-helper/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 小老師助手系統 API
  description: |
    Little Teacher Helper API - 讓小老師幫忙收回條和登記作業繳交狀況的系統
    
    ## 認證機制
    - 老師端: Session-based (Cookie)
    - 小老師端: 無需認證，透過 room code 加入
    
    ## 離線支援
    - 支援離線操作，資料暫存本機
    - 網路恢復後自動同步
  version: 1.0.0
  contact:
    name: Little Teacher Helper Team
servers:
  - url: /api
    description: Next.js API Routes
...
paths:
...
  /teachers:
    post:
      tags: [Teachers]
      summary: 建立老師帳號
      description: MVP 階段簡化版本，無需密碼驗證
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
      responses:
        '201':
          description: 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /submissions:
    patch:
      tags: [Submissions]
      summary: 更新繳交狀態
      description: 更新單筆或多筆繳交記錄
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionsRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                  submissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Submission'

  # ============= Sync =============
...
components:
  schemas:
...
    Teacher:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    CreateTeacherRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          format: email

    # ===== Room =====
...
    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        status:
          type: string
          enum: [SUBMITTED, NOT_SUBMITTED]
        updatedAt:
          type: string
          format: date-time
        syncedAt:
          type: string
          format: date-time
...
    UpdateSubmissionsRequest:
      type: object
      required: [submissions]
      properties:
        submissions:
          type: array
          items:
            type: object
            required: [studentId, itemId, status]
            properties:
              studentId:
                type: string
                format: uuid
              itemId:
                type: string
                format: uuid
              status:
                type: string
                enum: [SUBMITTED, NOT_SUBMITTED]

    # ===== Sync =====
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
  responses:
    BadRequest:
      description: 請求格式錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...